+++ src/global/constants/formTypeId.ts
@@ -1,0 +1,7 @@
+// TODO: would be nice to be able to pull this from the database on app start, once.
+export const formTypeId = {
+    team: 1,
+    "voyage member": 2,
+    user: 3,
+    meeting: 4,
+};